Handle DBus UnknownObject error when no collection is found
On Kubuntu 23.04, when trying to access a collection that doesn't
exist, it returns a UknownObject error instead of NoSuchObject.
Which makes me unable of installing the package on the system.

Handling this case, makes the KWallet dialog for creating the wallet
appear and the rest of workflow works as expected.